5547 - Split useLinksVerificationGuard
